api/requirements.txt
Flask==3.1.2
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.67.1
langchain-openai>=1.1.0
langchain>=0.3.0